Update all dependencies
This should fix following dependency resolver issue:

Dependencies could not be resolved because no versions of 'swift-identified-collections' match the requirement 0.3.2..<1.0.0 and root depends on 'swift-composable-architecture' 0.33.1..<1.0.0.
'swift-composable-architecture' >= 0.33.1 practically depends on 'swift-identified-collections' 0.3.2..<1.0.0 because 'swift-composable-architecture' 0.33.1 depends on 'swift-identified-collections' 0.3.2..<1.0.0.
'swift-composable-architecture' >= 0.33.2 practically depends on 'swift-identified-collections' 0.3.2..<1.0.0 because no versions of 'swift-composable-architecture' match the requirement {0.33.2..<0.34.0, 0.34.1..<1.0.0} and 'swift-composable-architecture' 0.34.0 depends on 'swift-identified-collections' 0.3.2..<1.0.0.